+++ src/database/operations/load/loadProposalBundle.ts
@@ -1,13 +1,22 @@
 import { loadBundle } from './loadBundle';
-import type { TinyPgParams } from 'tinypg';
 import type { JsonResultSet, Bundle, Proposal } from '../../../types';
 
-export const loadProposalBundle = async (
-	queryParameters: TinyPgParams,
-): Promise<Bundle<Proposal>> => {
+export const loadProposalBundle = async (queryParameters: {
+	offset: number;
+	limit: number;
+	search?: string;
+	organizationId?: number;
+}): Promise<Bundle<Proposal>> => {
+	const defaultQueryParameters = {
+		search: '',
+		organizationId: 0,
+	};
 	const bundle = await loadBundle<JsonResultSet<Proposal>>(
 		'proposals.selectWithPagination',
-		queryParameters,
+		{
+			...defaultQueryParameters,
+			...queryParameters,
+		},
 		'proposals',
 	);
 	const entries = bundle.entries.map((entry) => entry.object);